requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.9.0
qdrant-client>=1.7.0
google-generativeai>=0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
typer==0.9.0
sqlalchemy==2.0.23
asyncpg==0.29.0
trafilatura